Rudimentary suppport for drawing text to the framebuffer
We now have a `log` module that, through some really bad code, allows us
to write bitmaps, that are currently written in code (!), to the
framebuffer given to us by the bootloader.

This implementation will surely change in the feature, but this at least
allows for some limited form of printf debugging.

The goal is to eventually parse an image file containing a font atlas at
compile time and generate a packed (25 bits per character) array that
gets embedded into the kernel binary.

We also eventually want to add serial logging, meaning that the code
that currently resides in `log` will probably moved to `log::draw` or
something along those lines.

But this will do for now :)